refactor(itun): P4.2 rename code + route to Dashboard; Roster code → roster/ (ADR-021)
Completes P4 and the surface migration. Resolves the two-Dashboard collision by
coordinating a double rename so surface name, route, and code all agree.

- Roster (home surface): components/dashboard/ → components/roster/;
  Dashboard → Roster, DashboardSkeleton → RosterSkeleton, DashboardColumn →
  RosterColumn. Its user-facing name became "Roster" back in P0; the code now
  matches.
- Dashboard (play surface): components/play/ → components/dashboard/;
  PlayCockpit → Dashboard, Cockpit{Canvas,Chooser,Gauge} → Dashboard*,
  cockpit{Launch,Links}/playRules/play.css → dashboard*, ensureCockpitLinks →
  ensureDashboardLinks. "Play Cockpit"/"cockpit" surface prose → "Dashboard".
- Route: /play/$id → /dashboard/$id (routes/dashboard/$id.tsx); DashboardChooser
  navigation updated; routeTree.gen regenerated.
- Out of scope (left as-is): the playStateStore / workspace cockpitPrefs (live
  outside the two component dirs) and docs/.

Verified: bun run check:all fully green (lint, format, typecheck, all tests,
validate, knip). Executed as a git-mv rename set to preserve history.
