+++ apps/in-the-union-now/src/components/dashboard/__tests__/DashboardGauge.test.tsx
@@ -1,27 +1,27 @@
 /**
- * Unit tests for CockpitGauge — the bespoke dark instrument bar.
+ * Unit tests for DashboardGauge — the bespoke dark instrument bar.
  */
 
 import { describe, expect, test } from 'bun:test'
 import { render } from '@testing-library/react'
 
-import { CockpitGauge } from '../CockpitGauge'
+import { DashboardGauge } from '../DashboardGauge'
 
-describe('CockpitGauge', () => {
+describe('DashboardGauge', () => {
   test('renders max segments, fills up to value, shows value/max', () => {
-    const { container, getByText } = render(<CockpitGauge label="Heat" value={4} max={6} />)
+    const { container, getByText } = render(<DashboardGauge label="Heat" value={4} max={6} />)
     expect(container.querySelectorAll('.pc-seg').length).toBe(6)
     expect(container.querySelectorAll('.pc-seg.on').length).toBe(4)
     expect(getByText('4/6')).toBeTruthy()
   })
 
   test('filled segments at/after the danger index read as danger (redline)', () => {
-    const { container } = render(<CockpitGauge label="Heat" value={6} max={6} danger={4} />)
+    const { container } = render(<DashboardGauge label="Heat" value={6} max={6} danger={4} />)
     expect(container.querySelectorAll('.pc-seg.danger').length).toBe(2)
   })
 
   test('over-capacity values render extra segments honestly', () => {
-    const { container } = render(<CockpitGauge label="SP" value={8} max={6} />)
+    const { container } = render(<DashboardGauge label="SP" value={8} max={6} />)
     expect(container.querySelectorAll('.pc-seg').length).toBe(8)
     expect(container.querySelectorAll('.pc-seg.on').length).toBe(8)
   })
